- Fixed: The automap code had the check for rotation reversed. - Changed type PClass::FreeIndices to TArray<unsigned int> because that's   the type of the indices. - Fixed: makewad.c needs __cdecl for non-Windows builds. - Fixed: FinishThingdef didn't check whether the WeaponClass pointer in   AWeaponPiece was a valid name.
SVN r133 (trunk)

+++ src/dobject.cpp
@@ -343,7 +343,7 @@ CCMD (dumpclasses)
 }
 
 TArray<DObject *> DObject::Objects (TArray<DObject *>::NoInit);
-TArray<size_t> DObject::FreeIndices (TArray<size_t>::NoInit);
+TArray<unsigned int> DObject::FreeIndices (TArray<unsigned int>::NoInit);
 TArray<DObject *> DObject::ToDestroy (TArray<DObject *>::NoInit);
 bool DObject::Inactive;
